Investigate different API designs for repository
I believe the mutate API (add,delete,etc.) and LiveData observable design
is better than having the ViewModel and Repository editing a List for
a few reasons:
- Separation of command/query code for API consumers. You do lose
  the simplicity of just manipulating a list, however.
- Less code for consumers. No need to keep a reference to the List,
  modify it, then call postValue(). You just call add()/delete()/etc.
- No need to mock database. An in-memory database can be used here
  with immediate feedback from the observers that changes have been
  persisted.

NOTE: A good compromise could be having one update method that takes a
list and then using that to persist data and update the Repository. It
keeps the benefits listed above and allows the consumer to just work
with a normal list. It still has the problem of the consumer having to
manipulate a list, then call another method to keep their changes,
rather than just calling add/delete.